Add support for bindep.txt
We use bindep.txt files in zuul CI, to help identify which OS
dependencies are required for the job.  In this case, we add python26
support for fedora.

As we add more zuul jobs, this fix will expand to include more things.

Create tools/test-setup.sh to ensure pip is installed for python26.

Depends-On: https://github.com/ansible-network/ansible-zuul-jobs/pull/54
Signed-off-by: Paul Belanger <pabelanger@redhat.com>